hms_project_final/requirements.txt
django
psycopg2-binary
google-api-python-client
google-auth-oauthlib
requests
python-dotenv
protobuf